Separate server side code
