--- .babelrc
@@ -1,4 +1,0 @@
-{
-  "presets": ["@babel/preset-env", "@babel/preset-react"],
-  "plugins": ["@babel/transform-runtime", "styled-jsx/babel"]
-}
