--- index.js
@@ -1,3 +1,0 @@
-import { createExpressApp } from "./server/index.js";
-const app = createExpressApp();
-module.exports = app;
