+++ server/index.js
@@ -1,46 +1,12 @@
-import express from "express";
-import createExpressPino from "express-pino-logger";
-import bodyParser from "body-parser";
-import next from "next";
+import { createServer } from "./app";
+import { logger } from "./logger";
 
-import { insertMockData, setupDbConnection } from "./db";
-import { createApolloServer } from "./graphql/server";
-import { logger } from "../logger";
+const port = parseInt(process.env.PORT, 10) || 3000;
 
-const dev = process.env.NODE_ENV !== "production";
-const JWT_SECRET_KEY =
-  process.env.JWT_SECRET_KEY || "SHOULD_NOT_USE_THIS_IN_PRODUCTION";
-
-export async function createExpressApp() {
-  const nextApp = next({ dev, dir: "./client" });
-  const nextHandler = nextApp.getRequestHandler();
-  await nextApp.prepare();
-
-  const server = express();
-  const expressPino = createExpressPino({
-    logger
-  });
-  server.use(expressPino);
-  server.use(bodyParser.json());
-  server.use(bodyParser.urlencoded({ extended: true }));
-
-  const apolloServer = createApolloServer(JWT_SECRET_KEY);
-  apolloServer.applyMiddleware({ app: server });
-
-  // all requests to paths other than `/graphql` are processed by Nextjs
-  server.get("*", (req, res) => {
-    return nextHandler(req, res);
-  });
-
-  try {
-    await setupDbConnection(dev);
-    if (dev) {
-      insertMockData();
-    }
-  } catch (err) {
-    logger.error(err);
-    return;
-  }
-
-  return server;
+async function main() {
+  const server = await createServer();
+  const { url } = await server.listen(port);
+  logger.info(`> Ready on ${url}`);
 }
+
+main();
