+++ server/now.js
@@ -1,0 +1,3 @@
+import { createServer } from "./app.js";
+const app = createServer();
+module.exports = app;
