+++ server/package.json
@@ -1,0 +1,43 @@
+{
+  "name": "v8-bears-team-06",
+  "version": "0.0.1",
+  "description": "Add-project-description-here | Voyage-8 | https://chingu.io/",
+  "main": "index.js",
+  "scripts": {
+    "start:dev": "nodemon index.js --exec babel-node",
+    "lint": "eslint --ext jsx,js .",
+    "start": "node -r esm index.js"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/chingu-voyages/v8-bears-team-06.git"
+  },
+  "author": "",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/chingu-voyages/v8-bears-team-06/issues"
+  },
+  "homepage": "https://github.com/chingu-voyages/v8-bears-team-06#readme",
+  "dependencies": {
+    "@babel/cli": "^7.2.3",
+    "@babel/core": "^7.3.4",
+    "@babel/plugin-transform-runtime": "^7.3.4",
+    "@babel/preset-env": "^7.4.3",
+    "@babel/runtime": "^7.4.3",
+    "apollo-server": "^2.4.8",
+    "esm": "^3.2.22",
+    "graphql": "^14.1.1",
+    "jsonwebtoken": "^8.5.1",
+    "mongoose": "^5.4.19",
+    "pino": "^5.11.3"
+  },
+  "devDependencies": {
+    "@babel/node": "^7.2.2",
+    "babel-jest": "^24.5.0",
+    "eslint": "^5.16.0",
+    "eslint-config-prettier": "^4.1.0",
+    "mongodb-memory-server": "^4.2.1",
+    "nodemon": "^1.18.10",
+    "pino-pretty": "^2.6.0"
+  }
+}
